% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicPsychophysics.R
\name{GonzalezWuRightScaled}
\alias{GonzalezWuRightScaled}
\title{Create Predicted data patterns from a vanilla gonzalez & Wu model, scaled only on the right.  Suitable for one-sided bounds like line length}
\usage{
GonzalezWuRightScaled(stimuli, beta, delta, scaling = 1)
}
\arguments{
\item{stimuli}{a Vector of stimuli}

\item{delta}{The location of the prior / the elevation}

\item{gamma}{The curvature of the space, corresponding to a stevens beta. <1 yields overestimation of small values.}
}
\description{
This model calculates, given a set of parameters, the predicted results of an 'asymmetric' HD GW model. This 
model uses a central (but maybe off-center) reference point, and applies the G&W model from that central point ot each end
Each end may be separately scaled.
}
\keyword{proportionJudgments}
