% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicPsychophysics.R
\name{landyBrowerBiasedSpence}
\alias{landyBrowerBiasedSpence}
\title{landyBrowerBiasedSpence}
\usage{
landyBrowerBiasedSpence(stimuli, beta = 1, bias = 0, responses = NULL,
  tau = 1, mode = "prediction")
}
\arguments{
\item{stimuli}{a vector of stimuli, between 0 and 1}

\item{beta}{the power to raise the stimulus to, equivalent to the ratio of taus in the reconsturcted version}

\item{bias}{a fixed shift (in log odds space) of the stimuli}

\item{responses}{an optional vector of responses.}

\item{tau}{if log-likelihood is to be calculated, we must decide the predicted variability of responses. This is captured in tau, the precision}

\item{mode}{a mode. can validly be "prediction", or "logLikelihoodOfResponses". If the latter, responses must be given, and tau must be given}
}
\value{
A vector the transformed stimuli
}
\description{
landyBrowerBiasedSpence
}
\seealso{
classicGonzalezWu, spence
}
