% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicPsychophysics.R
\name{stevensPowerLaw}
\alias{stevensPowerLaw}
\title{stevensPowerLaw}
\usage{
stevensPowerLaw(stimuli, beta = 1, delta = 1, responses = NULL, tau = 1,
  mode = "prediction")
}
\arguments{
\item{stimuli}{a vector of stimuli, between 0 and inf}

\item{beta}{the power to raise the stimulus to, equivalent to the ratio of taus in the reconstucted version}

\item{delta}{proportionality constant--the constant multiple on the stimulus value}

\item{responses}{an optional vector of responses.}

\item{tau}{if log-likelihood is to be calculated, we must decide the predicted variability of responses. This is captured in tau, the precision}

\item{mode}{a mode. can validly be "prediction", or "logLikelihoodOfResponses". If the latter, responses must be given, and tau must be given}
}
\value{
A vector the transformed stimuli
}
\description{
stevensPowerLaw
}
\seealso{
gonzalezWu, spence
}
